ASSERT when launching the plugin process https://bugs.webkit.org/show_bug.cgi?id=46754 <rdar://problem/8484570>
Reviewed by Sam Weinig.

* UIProcess/Launcher/mac/ProcessLauncherMac.mm:
(WebKit::ProcessLauncher::launchProcess):
Don't assert if we fail to launch the process. Instead, set the connection identifier
and process ID to null and deallocate the port we created.

* UIProcess/Plugins/PluginProcessProxy.cpp:
(WebKit::PluginProcessProxy::pluginProcessCrashedOrFailedToLaunch):
Move all code from didClose here.

(WebKit::PluginProcessProxy::didClose):
Call pluginProcessCrashedOrFailedToLaunch.

(WebKit::PluginProcessProxy::didFinishLaunching):
If we failed to launch, call pluginProcessCrashedOrFailedToLaunch.

Canonical link: https://commits.webkit.org/59249@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@68562 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebKit2/UIProcess/Launcher/mac/ProcessLauncherMac.mm
@@ -130,13 +130,23 @@ void ProcessLauncher::launchProcess()
 
     pid_t processIdentifier;
     int result = posix_spawn(&processIdentifier, path, 0, &attr, (char *const*)args, *_NSGetEnviron());
-    ASSERT_UNUSED(result, !result);
 
     posix_spawnattr_destroy(&attr);
 
-    // Set up the termination notification handler and then ask the child process to continue.
-    setUpTerminationNotificationHandler(processIdentifier);
-    kill(processIdentifier, SIGCONT);
+    if (!result) {
+        // Set up the termination notification handler and then ask the child process to continue.
+        setUpTerminationNotificationHandler(processIdentifier);
+        kill(processIdentifier, SIGCONT);
+    } else {
+        // We failed to launch. Release the send right.
+        mach_port_deallocate(mach_task_self(), listeningPort);
+
+        // And the receive right.
+        mach_port_mod_refs(mach_task_self(), listeningPort, MACH_PORT_RIGHT_RECEIVE, -1);
+        
+        listeningPort = MACH_PORT_NULL;
+        processIdentifier = 0;
+    }
     
     // We've finished launching the process, message back to the main run loop.
     RunLoop::main()->scheduleWork(WorkItem::create(this, &ProcessLauncher::didFinishLaunchingProcess, processIdentifier, listeningPort));
